+++ app/Controllers/Home.php
@@ -105,7 +105,7 @@ class Home
         
         // ✅ 此时 app() 已可用！
 
-        dump(app()->getServiceIds()); // 查看所有服务 ID
+        //dump(app()->getServiceIds()); // 查看所有服务 ID
 
         // 日志测试
         // $logger = app('log');
